pyproject.toml
[tool.poetry]
name = "murkelhausen-app-v2"
version = "0.4.2"
description = ""
authors = ["Arkadius Schuchhardt <arkadius@schuchhardt.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
reflex = "^0.6.4"
pydantic = "^2.9.2"
requests = "^2.32.3"
cachetools = "^5.5.0"
babel = "^2.16.0"
beautifulsoup4 = "^4.12.3"
gcsa = "^2.5.1"
google-api-python-client = "^2.154.0"
oauth2client = "^4.1.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.7.3"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
